Move repl tests to radicle-repl from radicle-lang. Needed to rework having an Arbitrary module exported from radicle-lang. Unfortunately we're repeating the state world code in the radicle-lang for now.
